requirements.txt
transformers == 4.51.3
torch == 2.7.0
tiktoken==0.9.0
matplotlib==3.10.3
einops==0.8.1
transformers-stream-generator==0.0.5
accelerate==1.7.0
huggingface-hub==0.31.2
hf-xet == 1.1.1
librosa == 0.11.0
openai-whisper == 20240930
datasets == 3.6.0
python-dotenv == 1.1.0
openai == 1.79.0